guess we should include devices in the build
git-svn-id: http://svn.vuze.com/public/client@15796 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/platform/win32/access/impl/aedevice.cpp
@@ -165,8 +165,8 @@ JNIEXPORT jobject JNICALL Java_org_gudy_azureus2_platform_win32_access_impl_AEWi
 
 	if (bResult) 
 	{
-		ULONG diskSize = pdg.Cylinders.QuadPart * (ULONG)pdg.TracksPerCylinder *
-			(ULONG)pdg.SectorsPerTrack * (ULONG)pdg.BytesPerSector;
+		LONGLONG diskSize = pdg.Cylinders.QuadPart * pdg.TracksPerCylinder *
+			pdg.SectorsPerTrack * pdg.BytesPerSector;
 		addToMap(env, hashMap, methPut, clsLong, longInit, "MediaType", (jlong) pdg.MediaType);
 		addToMap(env, hashMap, methPut, clsLong, longInit, "DiskSize", (jlong) diskSize);
 	}
